+++ lookbook/previews/open_project/common/attribute_help_text_component_preview/in_a_form.html.erb
@@ -1,0 +1,26 @@
+<!--
+   The attribute help text is automatically rendered for each attribute in a form (when it exists).
+   The below example relies on the existing configuration of this instance.
+   Try to go to admin/attribute_help_texts and configure or delete an attribute help text for the fields below
+   -->
+
+<%
+  the_form = Class.new(ApplicationForm) do
+    form do |f|
+      f.text_field name: :version, label: attribute_name(:version), required: true, input_width: :small
+      f.text_field name: :assignee, label: attribute_name(:assignee), required: true, input_width: :medium
+    end
+  end
+
+  model = WorkPackage.new
+%>
+
+<%=
+  primer_form_with(
+    model:,
+    url: "/abc",
+    id: "asdf"
+  ) do |f|
+    render(the_form.new(f))
+  end
+%>
