+++ lookbook/previews/open_project/common/attribute_help_text_component_preview/playground.html.erb
@@ -1,0 +1,6 @@
+<!-- Note, that triggering the dialog does not work in Lookbook as the async dialog controller is not loaded here -->
+<%=
+  render OpenProject::Common::AttributeHelpTextComponent.new(help_text:) do |help_text|
+    help_text.with_additional_label { additional_text }
+  end
+%>
