--- lookbook/previews/open_project/common/attribute_help_text_component_preview/with_additional_label.html.erb
@@ -1,5 +1,0 @@
-<%=
-  render OpenProject::Common::AttributeHelpTextComponent.new(help_text:) do |help_text|
-    help_text.with_additional_label { "Show Attribute help text" }
-  end
-%>
